Remove roboto dependency from paper-styles
The roboto font is a dependency on a hard-coded CDN. Applications can
add the dependency and include the font as needed, rather than rely
paper-styles, which is used extensively throughout the Polymer
components. This is similar to other fonts listed in default styles
that might not be available on a system by default, but can be
resolved externally at the application level.

This change does not provide backward compatibility or a deprecation
period.